Change matrix type from double to int

+++ main.cpp
@@ -10,7 +10,7 @@ int main() {
     cout << "Ingrese cantidad de ecuaciones/incognitas: ";
     cin >> n;
 
-    vector<vector<double>> A(n, vector<double>(n));
+    vector<vector<int>> A(n, vector<int>(n));
     vector<int> b(n);
 
     cout << "Cargar matriz de coeficientes:\n";
@@ -36,7 +36,7 @@ int main() {
         return 0;
     }
 
-    vector<double> soluciones = resolverPorGaussJordan(A, b);
+    vector<int> soluciones = resolverPorGaussJordan(A, b);
 
     cout << "\nSoluciones:\n";
     for (int i = 0; i < soluciones.size(); i++) {
@@ -49,6 +49,7 @@ int main() {
 
     return 0;
 }
+
